switched all ids to usize

+++ crates/dslab-faas/src/controller.rs
@@ -35,7 +35,7 @@ impl Controller {
         }
     }
 
-    fn idle_deploy(&mut self, app_id: u64, time: f64) {
+    fn idle_deploy(&mut self, app_id: usize, time: f64) {
         let reg = self.function_registry.borrow();
         let app = reg.get_app(app_id).unwrap();
         if let Some(host) = self.idle_deployer.deploy(app, &self.hosts) {
