+++ crates/dslab-faas/src/invocation.rs
@@ -4,18 +4,18 @@ use crate::util::{Counter, VecMap};
 
 #[derive(Copy, Clone, Debug, Serialize)]
 pub struct InvocationRequest {
-    pub func_id: u64,
+    pub func_id: usize,
     pub duration: f64,
     pub time: f64,
-    pub id: u64,
+    pub id: usize,
 }
 
 #[derive(Copy, Clone)]
 pub struct Invocation {
-    pub id: u64,
+    pub id: usize,
     pub request: InvocationRequest,
-    pub host_id: u64,
-    pub container_id: u64,
+    pub host_id: usize,
+    pub container_id: usize,
     pub started: f64,
     pub finished: Option<f64>,
 }
@@ -27,7 +27,7 @@ pub struct InvocationRegistry {
 }
 
 impl InvocationRegistry {
-    pub fn new_invocation(&mut self, request: InvocationRequest, host_id: u64, container_id: u64, time: f64) {
+    pub fn new_invocation(&mut self, request: InvocationRequest, host_id: usize, container_id: usize, time: f64) {
         let id = request.id;
         let invocation = Invocation {
             id,
@@ -37,18 +37,18 @@ impl InvocationRegistry {
             started: time,
             finished: None,
         };
-        self.invocations.insert(id as usize, invocation);
+        self.invocations.insert(id, invocation);
     }
 
-    pub fn register_invocation(&mut self) -> u64 {
+    pub fn register_invocation(&mut self) -> usize {
         self.invocation_ctr.increment()
     }
 
-    pub fn get_invocation(&self, id: u64) -> Option<&Invocation> {
-        self.invocations.get(id as usize)
+    pub fn get_invocation(&self, id: usize) -> Option<&Invocation> {
+        self.invocations.get(id)
     }
 
-    pub fn get_invocation_mut(&mut self, id: u64) -> Option<&mut Invocation> {
-        self.invocations.get_mut(id as usize)
+    pub fn get_invocation_mut(&mut self, id: usize) -> Option<&mut Invocation> {
+        self.invocations.get_mut(id)
     }
 }
